+++ projects/demo/src/app/demo/pages/translate-page/translate-page.component.ts
@@ -1,7 +1,5 @@
 import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { TranslateService } from '@guillotinaweb/pastanaga-angular';
-import { filter } from 'rxjs/operators';
-import { ActivatedRoute, Router } from '@angular/router';
 
 @Component({
     selector: 'pa-translate-doc',
@@ -12,24 +10,33 @@ import { ActivatedRoute, Router } from '@angular/router';
 export class TranslatePageComponent {
     translations = `export const I18N_EN: TranslationEntries = {
     pastanaga: {
-        add: 'Add',
-        back: 'Back',
-        cancel: 'Cancel',
-        clear: 'Clear',
-        close: 'Close',
-        collapse: 'Collapse',
-        confirm: 'Confirm',
-        continue: 'Continue',
-        create: 'Create',
         datetime: {
             'a-few-seconds-ago': 'Just now',
             minutesAgo: '{{minutes}} mins ago',
             'one-minute-ago': '1 min ago',
             yesterday: 'Yesterday',
-            at: 'at',
         },
         // …
     },
+};
+`;
+    flattenTranslations = `export const I18N_EN: TranslationEntries = {
+  'pastanaga.datetime.a-few-seconds-ago': 'Just now',
+  'pastanaga.datetime.minutesAgo': '{{minutes}} mins ago',
+  'pastanaga.datetime.one-minute-ago': '1 min ago',
+  //…
+};
+`;
+    translationsMix = `export const I18N_EN: TranslationEntries = {
+  pastanaga: {
+    datetime: {
+      'a-few-seconds-ago': 'Just now',
+      minutesAgo: '{{minutes}} mins ago',
+    },
+  },
+  'pastanaga.datetime.one-minute-ago': '1 min ago',
+  'pastanaga.datetime.yesterday': 'Yesterday',
+  //…
 };
 `;
     appModule = `
@@ -52,21 +59,79 @@ import { DEMO_LA } from '../assets/i18n/la';
         PaTranslateModule.addTranslations([{'en_US': {...moreTranslationsForThisModule}}])
     ]`;
     directiveSimpleExample = `<span translate>demo-page.title</span>`;
-    directiveAttributeExample = `<span translate="demo-page.title"></span>`;
-    directiveWithParamsExample = `<span translate [translateParams]="{points: 10, total: 25}">demo-page.score</span>`;
+    directiveAttributeExample = `<span translate='demo-page.title'></span>`;
+    directiveWithParamsExample = `<span translate [translateParams]='{points: 10, total: 25}'>demo-page.score</span>`;
     pipeSimpleExample = `<span>{{ 'demo-page.title' | translate}}</span>`;
     pipeWithParamsExample = `<span>{{ 'demo-page.score' | translate:{points: 10, total: 25} }}</span>`;
+
+    ngxTranslateAndPastanaga = `@Component({
+  selector: 'app-root',
+  templateUrl: './app.component.html',
+  styleUrls: ['./app.component.scss']
+})
+export class AppComponent implements OnDestroy {
+  title = 'ngx-and-pa-translate';
+
+  private _terminator = new Subject<void>();
+
+  constructor(private ngxTranslate: ngxTranslateService, private paTranslate: paTranslateService) {
+    ngxTranslate.addLangs(['en', 'fr']);
+    ngxTranslate.setDefaultLang('en');
+
+    const browserLang = ngxTranslate.getBrowserLang() || '';
+    const lang = browserLang.match(/en|fr/) ? browserLang : 'en';
+    ngxTranslate.use(lang);
+    this.ngxTranslate.onLangChange.subscribe((event) => {
+      this.paTranslate.initTranslationsAndUse(event.lang,  event.translations);
+    });
+  }
+
+  ngOnDestroy() {
+    this._terminator.next();
+    this._terminator.complete();
+  }
+
+  changeLanguage() {
+    const lang = this.ngxTranslate.currentLang === 'en' ? 'fr' : 'en';
+    this.ngxTranslate.use(lang);
+  }
+}`;
+    noPaTranslateModule = `@NgModule({
+  declarations: [
+    AppComponent
+  ],
+  imports: [
+    AngularSvgIconModule.forRoot(),
+    BrowserModule,
+    HttpClientModule,
+    TranslateModule.forRoot({
+      loader: {
+        provide: TranslateLoader,
+        useFactory: HttpLoaderFactory,
+        deps: [HttpClient]
+      }
+    }),
+    PaDatePickerModule,
+    PaButtonModule,
+  ],
+  providers: [],
+  bootstrap: [AppComponent]
+})
+export class AppModule { }`;
+
+    dynamicallyChangePastanagaLang = `constructor(private translateService: TranslateService) {}
+
+updateLanguage(language: string) {
+    this.translateService.use(language);
+}`;
 
     languages = ['en', 'fr', 'latin'];
     currentLanguage = 'en';
 
-    constructor(private translateService: TranslateService, private activatedRoute: ActivatedRoute, private router: Router) {
-        this.activatedRoute.params.pipe(filter(params => !!params['la'])).subscribe(params => this.currentLanguage = params['la'])
-    }
+    constructor(private translateService: TranslateService) {}
 
     updateLanguage(language: string) {
         this.currentLanguage = language;
-        this.translateService.currentLanguage = language;
-        this.router.navigate(['/translate'], {queryParams: {la: language}});
+        this.translateService.use(language);
     }
 }