+++ projects/pastanaga-angular/src/lib/translate/translate.directive.spec.ts
@@ -1,7 +1,7 @@
 import { TranslateDirective } from './translate.directive';
-import { PA_TRANSLATIONS, TranslatePipe } from './translate.pipe';
+import { TranslatePipe } from './translate.pipe';
 import { ElementRef } from '@angular/core';
-import { PA_LANG, TranslateService } from './translate.service';
+import { PA_LANG, PA_TRANSLATIONS, TranslateService } from './translate.service';
 import { createDirectiveFactory, SpectatorDirective } from '@ngneat/spectator/jest';
 
 class MockElementRef implements ElementRef {
@@ -13,7 +13,14 @@ const obj2 = { common: 1, color: 'blue' };
 const obj3 = { common: 1, second: 3 };
 
 describe('TranslateDirective', () => {
-    const translations = { en_US: { hello: 'Bonjour', welcome: 'Bienvenue {{name}}' } };
+    const translations = {
+        en_US: {
+            hello: 'Bonjour',
+            welcome: 'Bienvenue {{name}}',
+            'pastanaga.cancel': 'Cancel',
+            calendar: { months: 'Mois', years: 'Années' },
+        },
+    };
 
     let directive: TranslateDirective;
 
@@ -38,6 +45,18 @@ describe('TranslateDirective', () => {
             expect(spectator.query('span')?.textContent).toBe(translations.en_US.hello);
         });
 
+        it('should translate a flatten key', () => {
+            spectator = createDirective(`<span translate>pastanaga.cancel</span>`);
+            spectator.detectChanges();
+            expect(spectator.query('span')?.textContent).toBe(translations.en_US['pastanaga.cancel']);
+        });
+
+        it('should translate a hierarchical key', () => {
+            spectator = createDirective(`<span translate>calendar.months</span>`);
+            spectator.detectChanges();
+            expect(spectator.query('span')?.textContent).toBe(translations.en_US.calendar.months);
+        });
+
         it('should translate a key with params', () => {
             spectator = createDirective(`<span translate [translateParams]="{name: 'Toto'}">welcome</span>`);
             spectator.detectChanges();
@@ -46,7 +65,9 @@ describe('TranslateDirective', () => {
 
         it('should translate a key with params and update the params when it changes', () => {
             let params = { name: 'Toto' };
-            spectator = createDirective(`<span translate [translateParams]="params">welcome</span>`, { hostProps: { params } });
+            spectator = createDirective(`<span translate [translateParams]="params">welcome</span>`, {
+                hostProps: { params },
+            });
             spectator.detectChanges();
             expect(spectator.query('span')?.textContent).toBe(translations.en_US.welcome.replace('{{name}}', 'Toto'));
 
@@ -69,12 +90,9 @@ describe('TranslateDirective', () => {
         });
     });
 
-    describe.skip('areEquals', () => {
+    describe('areEquals', () => {
         beforeEach(() => {
-            directive = new TranslateDirective(
-                new MockElementRef(),
-                new TranslatePipe(new TranslateService('en_US'), {}),
-            );
+            directive = new TranslateDirective(new MockElementRef(), new TranslateService('en_US', {}));
         });
 
         it('should be false when testing 2 objects with a different property ', () => {
@@ -90,5 +108,3 @@ describe('TranslateDirective', () => {
         });
     });
 });
-
-