+++ projects/pastanaga-angular/src/lib/translate/translate.model.ts
@@ -1,7 +1,14 @@
+export interface FlattenTranslationEntries {
+    [id: string]: string;
+}
 export interface TranslationEntries {
     [id: string]: string | TranslationEntries;
 }
 
 export interface Translation {
     [lang: string]: TranslationEntries;
+}
+
+export interface FlattenTranslation {
+    [lang: string]: FlattenTranslationEntries;
 }