+++ projects/pastanaga-angular/src/lib/translate/translate.module.ts
@@ -1,9 +1,10 @@
 import { ModuleWithProviders, NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { PA_TRANSLATIONS, TranslatePipe } from './translate.pipe';
+import { TranslatePipe } from './translate.pipe';
 import { TranslateDirective } from './translate.directive';
 import { Translation } from './translate.model';
 import { mergeTranslations } from './translate.utils';
+import { PA_TRANSLATIONS } from './translate.service';
 
 const _TRANSLATIONS: Translation = {};